[MOD] : 로그인, 회원가입으로 뷰모델 분리하기
